moving order requests to orders package as common resource
