+++ src/orders/order-request.ts
@@ -1,4 +1,4 @@
-import SearchResult from "./search/result";
+import SearchResult from "../archive/search/result";
 
 /**
  * @class OrderRequest wraps the details of an order request
